+++ R/dist_missing.R
@@ -27,14 +27,14 @@
 #'
 #' @export
 dist_missing <- function(length = 1) {
-  new_dist(x = rep(NA, length), class = "dist_na")
+  vctrs::vec_rep(NA_dist_, length)
 }
 
-# NA_dist_ <- dist_missing()
+NA_dist_ <- structure(list(NULL), class = c("distribution", "vctrs_vctr", "list"))
 
 #' @export
-format.dist_na <- function(x, ...){
-  rep_len("NA", length(x[["x"]]))
+format.dist_na <- function(x, ...) {
+  "NA"
 }
 
 #' @export
@@ -75,3 +75,13 @@ skewness.dist_na <- function(x, ...) NA_real_
 
 #' @export
 kurtosis.dist_na <- function(x, ...) NA_real_
+
+#' @export
+Math.dist_na <- function(x, ...) {
+  x
+}
+
+#' @export
+Ops.dist_na <- function(e1, e2) {
+  dist_missing(max(length(e1), length(e2)))
+}